feat: refactors shape rendering and adds hexagon and circles
The shape classes circle, triangle, rectangle and hexagon now uses a base
shape class which does all rendering. The inheritor of the base class
only specifies the vertices and indices for its specific shape.

Additionally, the CoreGraphicsContext is now a shared pointer and all
classes which depend on the CoreGraphicsContext for its destruction
now contain a shared pointer to this data.